backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
